mb/google/fizz: Enable VMX
We are enabling at the kernel level, but that is triggering an issue
where FSP expects it to be disabled so it forces a cold reboot on
every warm reboot, clearing the ramoops logs.  Enabling in BIOS so it
matches what the kernel expects.

This is the same change that were done for eve:
https://review.coreboot.org/#/c/22449/

BUG=None
BRANCH=None
TEST=echo PANIC >  /sys/kernel/debug/provoke-crash/DIRECT
     check for /dev/pstore/console-ramoops

Change-Id: Icd0bd01f5aee4c89f503eebba0808a1f3059e739
Signed-off-by: Shelley Chen <shchen@chromium.org>
Reviewed-on: https://review.coreboot.org/25251
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-by: Furquan Shaikh <furquan@google.com>